Move header files to their own directory
Move header files to "plustache" subdirectory

For example, when installing to /usr/local/include, header files will go in
/usr/local/include/plustache.

Also, add some details to the README:
 * Which header files to include in the example
 * How to build and run the unit tests
