+++ Makefile.am
@@ -1,7 +1,8 @@
+AM_CPPFLAGS = -I$(top_srcdir)/include
 LDFLAGS=-lboost_regex-mt
-include_HEADERS = include/template.hpp include/context.hpp include/plustache_types.hpp
+pkginclude_HEADERS = include/plustache/template.hpp include/plustache/context.hpp include/plustache/plustache_types.hpp
 check_PROGRAMS = test-program
-test_program_CPPFLAGS=-Ivendor/gtest-1.6.0/include
+test_program_CPPFLAGS=-Ivendor/gtest-1.6.0/include -I$(top_srcdir)/include
 test_program_LDFLAGS=-Lvendor/gtest-1.6.0/lib/.libs -lgtest -lboost_regex-mt
 
 test_program_SOURCES = tests/test_change_delimiter.cpp tests/test_collections_plustache.cpp tests/test_html_escape_plustache.cpp tests/test_inverted_sections.cpp tests/test_multiple_plustache.cpp tests/test_nested_sections_plustache.cpp tests/test_partials_plustache.cpp tests/test_sections_plustache.cpp tests/test_simple_plustache.cpp tests/tests.cpp
