+++ tests/test_change_delimiter.cpp
@@ -4,8 +4,8 @@
 #include <fstream>
 #include <gtest/gtest.h>
 
-#include "include/template.hpp"
-#include "include/plustache_types.hpp"
+#include <plustache/template.hpp>
+#include <plustache/plustache_types.hpp>
 
 // The fixture for testing class Foo.
 class ChangeDelimiterTest : public ::testing::Test
