+++ tests/test_collections_plustache.cpp
@@ -3,9 +3,9 @@
 #include <fstream>
 #include <gtest/gtest.h>
 
-#include "include/template.hpp"
-#include "include/context.hpp"
-#include "include/plustache_types.hpp"
+#include <plustache/template.hpp>
+#include <plustache/context.hpp>
+#include <plustache/plustache_types.hpp>
 
 // The fixture for testing class Foo.
 class CollectionsTest : public ::testing::Test
