+++ tests/test_partials_plustache.cpp
@@ -3,8 +3,8 @@
 #include <fstream>
 #include <gtest/gtest.h>
 
-#include "include/template.hpp"
-#include "include/plustache_types.hpp"
+#include <plustache/template.hpp>
+#include <plustache/plustache_types.hpp>
 
 // The fixture for testing class Foo.
 class PartialsTest : public ::testing::Test
